+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Added
 
 - new option renderOverlappingSchematics: render every schematic placement in areas where their bounds overlap, instead of only one of them
+- new hotkey layerChangeAmountModifier: hold it and scroll the mouse wheel to raise/lower layerChangeAmount in game
 
 ## Fixed